feat(mcp): expose embedding preflight
Problem: #1503 made embedding readiness visible to MCP clients, but agents still had no read-only way to ask what a bounded semantic catch-up window would cost without scraping `polylogue embed preflight` output.

What changed: move embedding preflight construction into `polylogue.storage.embeddings.preflight`, keep the CLI wrappers over that shared contract, and add an MCP `embedding_preflight` tool that returns the canonical machine-readable preflight payload. Refresh the MCP schema witness, generated topology/catalog docs, and search documentation.

Verification: focused MCP/CLI tests -> 62 passed; mypy on touched files -> success; devtools verify --json -> exit_code 0.

Ref #1503
