+++ docs/topology-status.md
@@ -29,10 +29,10 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 720
+- **Stable** (no move scoped): 721
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 16
 - **TBD** (cell needs explicit assignment): 0
-- **Total declared**: 853
-- **Realized polylogue/**/*.py**: 853 files declared
+- **Total declared**: 854
+- **Realized polylogue/**/*.py**: 854 files declared
 <!-- END GENERATED: topology-status -->
