+++ docs/verification-catalog.md
@@ -78,12 +78,12 @@ This catalog is generated from the proof-obligation kernel. It records subjects,
 | `polylogue diagnostics turns` | `polylogue/cli/commands/diagnostics.py:92` `polylogue.cli.commands.diagnostics.turns_command` |
 | `polylogue doctor` | `polylogue/cli/click_app.py` `polylogue doctor` |
 | `polylogue embed` | `polylogue/cli/click_app.py` `polylogue embed` |
-| `polylogue embed activate` | `polylogue/cli/commands/embed.py:439` `polylogue.cli.commands.embed.activate_subcommand` |
-| `polylogue embed backfill` | `polylogue/cli/commands/embed.py:520` `polylogue.cli.commands.embed.backfill_subcommand` |
-| `polylogue embed disable` | `polylogue/cli/commands/embed.py:446` `polylogue.cli.commands.embed.disable_subcommand` |
-| `polylogue embed enable` | `polylogue/cli/commands/embed.py:378` `polylogue.cli.commands.embed.enable_subcommand` |
-| `polylogue embed preflight` | `polylogue/cli/commands/embed.py:461` `polylogue.cli.commands.embed.preflight_subcommand` |
-| `polylogue embed status` | `polylogue/cli/commands/embed.py:732` `polylogue.cli.commands.embed.status_subcommand` |
+| `polylogue embed activate` | `polylogue/cli/commands/embed.py:276` `polylogue.cli.commands.embed.activate_subcommand` |
+| `polylogue embed backfill` | `polylogue/cli/commands/embed.py:357` `polylogue.cli.commands.embed.backfill_subcommand` |
+| `polylogue embed disable` | `polylogue/cli/commands/embed.py:283` `polylogue.cli.commands.embed.disable_subcommand` |
+| `polylogue embed enable` | `polylogue/cli/commands/embed.py:215` `polylogue.cli.commands.embed.enable_subcommand` |
+| `polylogue embed preflight` | `polylogue/cli/commands/embed.py:298` `polylogue.cli.commands.embed.preflight_subcommand` |
+| `polylogue embed status` | `polylogue/cli/commands/embed.py:569` `polylogue.cli.commands.embed.status_subcommand` |
 | `polylogue export` | `polylogue/cli/click_app.py` `polylogue export` |
 | `polylogue facets` | `polylogue/cli/click_app.py` `polylogue facets` |
 | `polylogue feedback` | `polylogue/cli/click_app.py` `polylogue feedback` |
