+++ tests/unit/mcp/test_envelope_contracts.py
@@ -81,6 +81,7 @@ TOOL_CONTRACT: dict[str, ToolKind] = {
     "cost_outlook": "single_object",
     "stats": "single_object",
     "embedding_status": "single_object",
+    "embedding_preflight": "single_object",
     "facets": "single_object",
     "build_context_pack": "single_object",
     "readiness_check": "single_object",
